+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerRequired.test.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import DatePicker from '../../../../lib/DatePicker';
+
+const DatePickerRequired = () => (
+  <DatePicker
+    name="date-input"
+    required
+  />
+);
+
+export default DatePickerRequired;