+++ packages/terra-date-picker/tests/jest/DatePicker.test.jsx
@@ -147,3 +147,11 @@ it('should render a disabled date picker', () => {
     </IntlProvider>));
   expect(datePicker).toMatchSnapshot();
 });
+
+it('should render a required date input and date picker', () => {
+  const datePicker = shallow((
+    <IntlProvider locale={locale} messages={messages}>
+      <DatePicker name="date-input" utcOffset={0} required />
+    </IntlProvider>));
+  expect(datePicker).toMatchSnapshot();
+});